Rename test app package hotadd -> telemetry
These tests don't seem to involve hot adding anything
